wip: split inventory recommended apps surface
