+++ packages/gui-web/src/InventorySurfaces.tsx
@@ -1,12 +1,13 @@
 import type { GuiAppActionId, GuiAppActionJobSummary, GuiManagedAppSummary, GuiResponseEnvelope, GuiStatusData } from "@aidevops/gui-shared";
-import { type Dispatch, type MouseEvent as ReactMouseEvent, type ReactElement, type SetStateAction, useEffect, useRef, useState } from "react";
+import { type Dispatch, type ReactElement, type SetStateAction, useEffect, useRef, useState } from "react";
 import type { IconType } from "react-icons";
-import { FaApple, FaLinux, FaWindows } from "react-icons/fa";
-import { FiChevronDown, FiCode, FiDownload, FiExternalLink, FiGlobe, FiMonitor, FiRefreshCw, FiRepeat, FiTerminal, FiTrash2 } from "react-icons/fi";
-import { IoLogoAndroid } from "react-icons/io";
-import { SiIos } from "react-icons/si";
+import { FiChevronDown, FiDownload, FiRefreshCw, FiRepeat, FiTrash2 } from "react-icons/fi";
 import type { InventoryColumn } from "./app-model";
 import { installationRows, text } from "./app-model";
+import { TabNav, type TabOption } from "./AppTabs";
+import { OriginLink } from "./ExternalLinks";
+import { RecommendedAppsSurface } from "./RecommendedAppsSurface";
+import { recommendedAppMatchesFilters, recommendedApps, type RecommendedOsId, type RecommendedPlatformFilterId } from "./recommended-apps";
 
 interface DraftInventoryRow {
   id: string;
@@ -137,14 +138,6 @@ function SummaryChip({ label, value }: { label: string; value: string }): ReactE
   return <span className="managed-summary-chip"><small>{label}</small><span>{value}</span></span>;
 }
 
-function OriginLink({ href, label }: { href: string; label: string }): ReactElement {
-  if (href.length === 0) {
-    return <span className="origin-missing">{label}: source pending</span>;
-  }
-
-  return <a href={href} onClick={(event) => openExternalLink(event, href)} rel="noreferrer" target="_blank" title={href}>{label} <FiExternalLink aria-hidden="true" /></a>;
-}
-
 function ToggleSwitch({ checked, disabled = false, label, onChange }: { checked: boolean; disabled?: boolean; label: string; onChange: (checked: boolean) => void }): ReactElement {
   return <button aria-pressed={checked} className={checked ? "managed-toggle checked" : "managed-toggle"} data-tooltip={disabled ? "Essential aidevops component; policy is locked on" : undefined} disabled={disabled} onClick={() => onChange(!checked)} title={disabled ? "Essential aidevops component; policy is locked on" : undefined} type="button"><span aria-hidden="true" className={checked ? "switch-track checked" : "switch-track"}><span /></span>{label}</button>;
 }
@@ -158,7 +151,7 @@ function AppLogLink({ job }: { job: GuiAppActionJobSummary }): ReactElement {
 }
 
 function AppActionButton({ action, app, commandPreview, disabled, onJob }: { action: GuiAppActionId; app: GuiManagedAppSummary; commandPreview: string; disabled: boolean; onJob: (job: GuiAppActionJobSummary) => void }): ReactElement {
-  const icon = action === "install" ? <FiDownload /> : action === "update" ? <FiRefreshCw /> : action === "reinstall" ? <FiRepeat /> : <FiTrash2 />;
+  const ActionIcon = appActionIcons[action];
   const [confirmOpen, setConfirmOpen] = useState(false);
 
   async function runAction(): Promise<void> {
@@ -172,7 +165,7 @@ function AppActionButton({ action, app, commandPreview, disabled, onJob }: { act
   }
 
   return <>
-    <button aria-label={`${action} ${app.name}`} className={action === "remove" ? "app-action-button remove" : "app-action-button"} data-tooltip={commandPreview} disabled={disabled} onClick={() => setConfirmOpen(true)} type="button">{icon}<span>{action}</span></button>
+    <button aria-label={`${action} ${app.name}`} className={action === "remove" ? "app-action-button remove" : "app-action-button"} data-tooltip={commandPreview} disabled={disabled} onClick={() => setConfirmOpen(true)} type="button"><ActionIcon /><span>{action}</span></button>
     {confirmOpen ? <ConfirmActionModal action={action} app={app} commandPreview={commandPreview} close={() => setConfirmOpen(false)} confirm={() => { setConfirmOpen(false); void runAction(); }} /> : null}
   </>;
 }
@@ -240,55 +233,44 @@ function sortedManagedApps(apps: GuiManagedAppSummary[]): GuiManagedAppSummary[]
 }
 
 function managedCategoryForApp(app: GuiManagedAppSummary): ManagedCategoryId {
-  if (["core", "safety", "automation"].includes(app.category)) {
-    return "core";
-  }
-  if (app.category.includes("cli")) {
-    return "cli";
-  }
-  if (app.category.includes("ai")) {
-    return "ai";
-  }
-  if (["desktop", "editor"].includes(app.category)) {
-    return "desktop";
-  }
-  if (["terminal"].includes(app.category) || app.name.toLowerCase().includes("terminal")) {
-    return "terminal";
-  }
+  const matchedCategory = managedCategoryRules.find((rule) => rule.matches(app));
 
-  return "cli";
+  return matchedCategory?.id ?? "cli";
 }
 
 function isEssentialManagedApp(app: GuiManagedAppSummary): boolean {
-  return ["core", "safety", "automation"].includes(app.category);
+  return categoryIsOneOf(app.category, coreManagedAppCategories);
 }
 
 type AppCollectionId = "aidevops" | "recommended";
 type ManagedCategoryId = "core" | "ai" | "cli" | "desktop" | "terminal";
-type RecommendedOsId = "all" | "macos" | "linux" | "windows" | "ios" | "android";
-type RecommendedPlatformId = "webapp" | "saas" | "cli" | "api";
-type RecommendedPlatformFilterId = "all" | RecommendedPlatformId;
 type ManagedPolicyId = "setup" | "update";
 type ManagedPolicyToggleState = Record<string, Partial<Record<ManagedPolicyId, boolean>>>;
 
-interface TabOption<T extends string> {
-  id: T;
-  label: string;
+interface ManagedCategoryRule {
+  id: ManagedCategoryId;
+  matches: (app: GuiManagedAppSummary) => boolean;
 }
 
-interface RecommendedApp {
-  name: string;
-  description: string;
-  websiteUrl: string;
-  alternativeToUrl?: string;
-  repoUrl?: string;
-  iosUrl?: string;
-  androidUrl?: string;
-  fdroidUrl?: string;
-  obtainiumUrl?: string;
-  os: RecommendedOsId[];
-  platforms: RecommendedPlatformId[];
-}
+const appActionIcons: Record<GuiAppActionId, IconType> = {
+  install: FiDownload,
+  update: FiRefreshCw,
+  reinstall: FiRepeat,
+  remove: FiTrash2,
+};
+
+const coreManagedAppCategories = ["core", "safety", "automation"];
+const desktopManagedAppCategories = ["desktop", "editor"];
+const terminalManagedAppCategories = ["terminal"];
+
+const managedCategoryRules: ManagedCategoryRule[] = [
+  { id: "core", matches: (app) => categoryIsOneOf(app.category, coreManagedAppCategories) },
+  { id: "cli", matches: (app) => app.category.includes("cli") },
+  { id: "ai", matches: (app) => app.category.includes("ai") },
+  { id: "desktop", matches: (app) => categoryIsOneOf(app.category, desktopManagedAppCategories) },
+  { id: "terminal", matches: (app) => categoryIsOneOf(app.category, terminalManagedAppCategories) },
+  { id: "terminal", matches: (app) => app.name.toLowerCase().includes("terminal") },
+];
 
 const appCollectionTabs: TabOption<AppCollectionId>[] = [
   { id: "aidevops", label: "aidevops" },
@@ -303,160 +285,8 @@ const managedCategoryTabs: TabOption<ManagedCategoryId>[] = [
   { id: "terminal", label: "Terminal" },
 ];
 
-const recommendedPlatformTabs: TabOption<RecommendedPlatformFilterId>[] = [
-  { id: "all", label: "All" },
-  { id: "webapp", label: "WebApp" },
-  { id: "saas", label: "SaaS" },
-  { id: "cli", label: "CLI" },
-  { id: "api", label: "API" },
-];
-
-const recommendedOsTabs: TabOption<RecommendedOsId>[] = [
-  { id: "all", label: "All OS" },
-  { id: "macos", label: "macOS" },
-  { id: "linux", label: "Linux" },
-  { id: "windows", label: "Windows" },
-  { id: "ios", label: "iOS" },
-  { id: "android", label: "Android" },
-];
-
-const recommendedApps = ([
-  { name: "Affinity Studio", description: "Creative design suite.", websiteUrl: "https://www.affinity.studio/", alternativeToUrl: "https://alternativeto.net/software/affinity-1/", os: ["macos", "windows", "ios"], platforms: [] },
-  { name: "Bitwarden", description: "Open source password manager.", websiteUrl: "https://bitwarden.com", alternativeToUrl: "https://alternativeto.net/software/bitwarden--free-password-manager/", repoUrl: "https://github.com/bitwarden", os: ["macos", "linux", "windows", "ios", "android"], platforms: ["webapp", "saas", "cli", "api"] },
-  { name: "Brave Browser", description: "Privacy-focused web browser.", websiteUrl: "https://brave.com/", alternativeToUrl: "https://alternativeto.net/software/brave/", os: ["macos", "linux", "windows", "ios", "android"], platforms: [] },
-  { name: "Cloudron", description: "Self-hosted app platform.", websiteUrl: "https://www.cloudron.io/", alternativeToUrl: "https://alternativeto.net/software/cloudron/", os: ["linux"], platforms: ["webapp", "cli", "api"] },
-  { name: "Collabora Online", description: "Online office collaboration.", websiteUrl: "https://www.collaboraonline.com/collabora-online/", alternativeToUrl: "https://alternativeto.net/software/collabora-online/", os: ["linux"], platforms: ["webapp", "api"] },
-  { name: "Cometly", description: "Marketing attribution analytics.", websiteUrl: "https://www.cometly.com/", os: [], platforms: ["saas", "api"] },
-  { name: "DaVinci Resolve", description: "Professional video editing, color, VFX, and audio post-production.", websiteUrl: "https://www.blackmagicdesign.com/products/davinciresolve/", alternativeToUrl: "https://alternativeto.net/software/davinci-resolve/", os: ["macos", "linux", "windows"], platforms: [] },
-  { name: "DocuSeal", description: "Open source document signing.", websiteUrl: "https://www.docuseal.com/", alternativeToUrl: "https://alternativeto.net/software/docuseal/", os: ["linux"], platforms: ["webapp", "saas", "api"] },
-  { name: "Element", description: "Matrix messaging client.", websiteUrl: "https://element.io/", alternativeToUrl: "https://alternativeto.net/software/element-app/", os: ["macos", "linux", "windows", "ios", "android"], platforms: ["webapp"] },
-  { name: "Enpass", description: "Offline-first password manager.", websiteUrl: "https://www.enpass.io/", alternativeToUrl: "https://alternativeto.net/software/enpass/", os: ["macos", "linux", "windows", "ios", "android"], platforms: [] },
-  { name: "EspoCRM", description: "Open source CRM.", websiteUrl: "https://www.espocrm.com/", alternativeToUrl: "https://alternativeto.net/software/espocrm/", os: ["linux"], platforms: ["webapp", "saas", "api"] },
-  { name: "Fathom Analytics", description: "Privacy-first analytics.", websiteUrl: "https://usefathom.com/", alternativeToUrl: "https://alternativeto.net/software/fathom-analytics/", os: [], platforms: ["saas", "api"] },
-  { name: "FontBase", description: "Font management and reference tool.", websiteUrl: "https://fontba.se/", alternativeToUrl: "https://alternativeto.net/software/fontbase/", os: ["macos"], platforms: [] },
-  { name: "Forgejo", description: "Self-hosted Git forge.", websiteUrl: "https://forgejo.org/", alternativeToUrl: "https://alternativeto.net/software/forgejo/", os: ["linux"], platforms: ["webapp", "cli", "api"] },
-  { name: "Ghost", description: "Publishing platform.", websiteUrl: "https://ghost.org/", alternativeToUrl: "https://alternativeto.net/software/ghost/", os: ["linux"], platforms: ["webapp", "saas", "cli", "api"] },
-  { name: "Gitea", description: "Self-hosted Git service.", websiteUrl: "https://about.gitea.com/", alternativeToUrl: "https://alternativeto.net/software/gitea/", os: ["linux", "windows"], platforms: ["webapp", "cli", "api"] },
-  { name: "GitLab", description: "DevSecOps and Git hosting platform.", websiteUrl: "https://gitlab.com/", alternativeToUrl: "https://alternativeto.net/software/gitlab/", os: ["linux"], platforms: ["webapp", "saas", "cli", "api"] },
-  { name: "LibreOffice", description: "Open source office suite.", websiteUrl: "https://www.libreoffice.org/", alternativeToUrl: "https://alternativeto.net/software/libreoffice/", os: ["macos", "linux", "windows"], platforms: [] },
-  { name: "LocalWP", description: "Local WordPress development.", websiteUrl: "https://localwp.com/", alternativeToUrl: "https://alternativeto.net/software/local-by-flywheel/", os: ["macos", "linux", "windows"], platforms: [] },
-  { name: "Matomo", description: "Open analytics platform.", websiteUrl: "https://matomo.org/", alternativeToUrl: "https://alternativeto.net/software/piwik/", os: ["linux"], platforms: ["webapp", "saas", "api"] },
-  { name: "Nextcloud", description: "Open source content collaboration platform.", websiteUrl: "https://nextcloud.com/", alternativeToUrl: "https://alternativeto.net/software/nextcloud/", repoUrl: "https://github.com/nextcloud", os: ["macos", "linux", "windows", "ios", "android"], platforms: ["webapp", "saas", "api"] },
-  { name: "Nextcloud Talk", description: "Open source video calls, chat, and collaboration for Nextcloud.", websiteUrl: "https://nextcloud.com/talk/", repoUrl: "https://github.com/nextcloud/spreed", iosUrl: "https://apps.apple.com/us/app/nextcloud-talk/id1296825574", androidUrl: "https://play.google.com/store/apps/details?id=com.nextcloud.talk2&hl=en", os: ["ios", "android"], platforms: ["webapp"] },
-  { name: "OBS Studio", description: "Open source video recording and live streaming.", websiteUrl: "https://obsproject.com/", alternativeToUrl: "https://alternativeto.net/software/open-broadcaster-software/", repoUrl: "https://github.com/obsproject/obs-studio", os: ["macos", "linux", "windows"], platforms: [] },
-  { name: "ONLYOFFICE", description: "Office and document collaboration.", websiteUrl: "https://www.onlyoffice.com/", alternativeToUrl: "https://alternativeto.net/software/onlyoffice/", repoUrl: "https://github.com/ONLYOFFICE/", os: ["macos", "linux", "windows", "ios", "android"], platforms: ["webapp", "saas", "api"] },
-  { name: "OpenScreen", description: "Open screen-sharing project.", websiteUrl: "https://github.com/getopenscreen/openscreen/releases", alternativeToUrl: "https://alternativeto.net/software/openscreen/", repoUrl: "https://github.com/getopenscreen/openscreen", os: ["macos", "linux", "windows"], platforms: [] },
-  { name: "Osaurus", description: "AI app workspace.", websiteUrl: "https://osaurus.ai/", alternativeToUrl: "https://alternativeto.net/software/osaurus/", os: ["macos"], platforms: [] },
-  { name: "Parallels Desktop", description: "Run Windows, Linux, and virtual machines on macOS.", websiteUrl: "https://www.parallels.com/products/desktop/", alternativeToUrl: "https://alternativeto.net/software/parallels-desktop/about/", os: ["macos"], platforms: [] },
-  { name: "PDF Studio", description: "PDF editor.", websiteUrl: "https://www.qoppa.com/pdfstudio/", alternativeToUrl: "https://alternativeto.net/software/qoppa-pdf-studio/", os: ["macos", "linux", "windows"], platforms: [] },
-  { name: "Pixelmator Pro", description: "Professional image editor for macOS.", websiteUrl: "https://www.apple.com/pixelmator-pro/", os: ["macos"], platforms: [] },
-  { name: "PostHog", description: "Product analytics platform.", websiteUrl: "https://posthog.com/", alternativeToUrl: "https://alternativeto.net/software/posthog/", os: ["linux"], platforms: ["webapp", "saas", "cli", "api"] },
-  { name: "Postiz", description: "Social media scheduling.", websiteUrl: "https://postiz.com/", alternativeToUrl: "https://alternativeto.net/software/postiz/", os: ["linux"], platforms: ["webapp", "saas", "api"] },
-  { name: "PrivateBin", description: "Zero-knowledge pastebin.", websiteUrl: "https://privatebin.info/", alternativeToUrl: "https://alternativeto.net/software/privatebin/", repoUrl: "https://github.com/PrivateBin/PrivateBin", os: ["linux"], platforms: ["webapp", "api"] },
-  { name: "Proxmox", description: "Virtualization platform.", websiteUrl: "https://www.proxmox.com/", alternativeToUrl: "https://alternativeto.net/software/proxmox-virtual-environment/", os: ["linux"], platforms: ["webapp", "cli", "api"] },
-  { name: "QuickFile", description: "Accounting platform.", websiteUrl: "https://www.quickfile.co.uk/", os: [], platforms: ["saas", "api"] },
-  { name: "Reframed", description: "Creative framing utility.", websiteUrl: "https://www.reframed.dev/", alternativeToUrl: "https://alternativeto.net/software/reframed/", os: ["macos"], platforms: [] },
-  { name: "Rybbit", description: "Web analytics.", websiteUrl: "https://rybbit.com/", alternativeToUrl: "https://alternativeto.net/software/rybbit/", os: ["linux"], platforms: ["webapp", "saas", "api"] },
-  { name: "SchildiChat", description: "Matrix messaging client.", websiteUrl: "https://schildi.chat/", alternativeToUrl: "https://alternativeto.net/software/schildichat/", os: ["macos", "linux", "windows", "android"], platforms: ["webapp"] },
-  { name: "ScreenFlow", description: "Screen recording and editing.", websiteUrl: "https://www.telestream.net/screenflow/", alternativeToUrl: "https://alternativeto.net/software/screenflow/", os: ["macos", "ios"], platforms: [] },
-  { name: "SEO Utils", description: "Desktop SEO tools.", websiteUrl: "https://seoutils.app/", os: ["macos", "linux", "windows"], platforms: [] },
-  { name: "Shottr", description: "macOS screenshot utility.", websiteUrl: "https://shottr.cc/", alternativeToUrl: "https://alternativeto.net/software/shottr/", os: ["macos"], platforms: [] },
-  { name: "Signal", description: "Private messenger.", websiteUrl: "https://signal.org/", alternativeToUrl: "https://alternativeto.net/software/signal-private-messenger/", os: ["macos", "linux", "windows", "ios", "android"], platforms: [] },
-  { name: "SimpleX Chat", description: "Private messenger with no user IDs.", websiteUrl: "https://simplex.chat/", alternativeToUrl: "https://alternativeto.net/software/simplex-chat/about/", repoUrl: "https://github.com/simplex-chat", os: ["macos", "linux", "windows", "ios", "android"], platforms: ["cli"] },
-  { name: "Telegram", description: "Cloud-based mobile and desktop messaging app.", websiteUrl: "https://telegram.org/", iosUrl: "https://telegram.org/dl/ios", androidUrl: "https://telegram.org/android", os: ["macos", "linux", "windows", "ios", "android"], platforms: ["webapp", "api"] },
-  { name: "Thunderbird", description: "Email and calendar app.", websiteUrl: "https://www.thunderbird.net/", alternativeToUrl: "https://alternativeto.net/software/mozilla-thunderbird/about/", os: ["macos", "linux", "windows"], platforms: [] },
-  { name: "Ubicloud", description: "Open cloud platform.", websiteUrl: "https://www.ubicloud.com/", alternativeToUrl: "https://alternativeto.net/software/ubicloud/about/", repoUrl: "https://github.com/ubicloud/ubicloud", os: ["linux"], platforms: ["webapp", "saas", "cli", "api"] },
-  { name: "Vaultwarden", description: "Alternative Bitwarden server implementation.", websiteUrl: "https://github.com/dani-garcia/vaultwarden", alternativeToUrl: "https://alternativeto.net/software/vaultwarden/", repoUrl: "https://github.com/dani-garcia/vaultwarden", os: ["linux"], platforms: ["webapp", "api"] },
-  { name: "VideoProc", description: "Video processing toolkit.", websiteUrl: "https://www.videoproc.com/", alternativeToUrl: "https://alternativeto.net/software/videoproc/about/", os: ["macos", "windows"], platforms: [] },
-  { name: "VirtualBox", description: "Virtualization app.", websiteUrl: "https://www.virtualbox.org/", alternativeToUrl: "https://alternativeto.net/software/virtualbox/about/", os: ["macos", "linux", "windows"], platforms: ["cli"] },
-  { name: "WordPress", description: "Open source publishing platform.", websiteUrl: "https://wordpress.org/", alternativeToUrl: "https://alternativeto.net/software/wordpress/about/", os: ["linux"], platforms: ["webapp", "saas", "cli", "api"] },
-] satisfies RecommendedApp[]).sort((left, right) => left.name.localeCompare(right.name, undefined, { sensitivity: "base" }));
-
-function TabNav<T extends string>({ label, onChange, tabs, value }: { label: string; onChange: (value: T) => void; tabs: TabOption<T>[]; value: T }): ReactElement {
-  return <div aria-label={label} className="pill-tabs app-subnav" role="tablist">{tabs.map((tab) => <button aria-selected={value === tab.id} className={value === tab.id ? "active" : ""} key={tab.id} onClick={() => onChange(tab.id)} role="tab" type="button">{tab.label}</button>)}</div>;
-}
-
-function RecommendedAppsSurface({ apps, os, platform, setOs, setPlatform }: { apps: RecommendedApp[]; os: RecommendedOsId; platform: RecommendedPlatformFilterId; setOs: (os: RecommendedOsId) => void; setPlatform: (platform: RecommendedPlatformFilterId) => void }): ReactElement {
-  return <>
-    <div className="recommended-filter-tabs">
-      <TabNav label="Recommended app platform filters" tabs={recommendedPlatformTabs} value={platform} onChange={setPlatform} />
-      <TabNav label="Recommended app operating system filters" tabs={recommendedOsTabs} value={os} onChange={setOs} />
-    </div>
-    <div className="recommended-app-grid">
-      {apps.map((app) => <RecommendedAppCard app={app} key={app.name} setOs={setOs} setPlatform={setPlatform} />)}
-    </div>
-  </>;
-}
-
-function RecommendedAppCard({ app, setOs, setPlatform }: { app: RecommendedApp; setOs: (os: RecommendedOsId) => void; setPlatform: (platform: RecommendedPlatformFilterId) => void }): ReactElement {
-  return <article className="recommended-app-card">
-    <div>
-      <strong>{app.name}</strong>
-      <p>{app.description}</p>
-    </div>
-    <div className="app-icon-groups">
-      <PlatformIconList platforms={app.platforms} setPlatform={setPlatform} />
-      <OsIconList os={app.os} setOs={setOs} />
-    </div>
-    <div className="managed-app-links plain-links">
-      <OriginLink href={app.websiteUrl} label={text.website} />
-      {app.alternativeToUrl ? <OriginLink href={app.alternativeToUrl} label="AlternativeTo" /> : null}
-      {app.repoUrl ? <OriginLink href={app.repoUrl} label="Repo" /> : null}
-      {app.iosUrl ? <OriginLink href={app.iosUrl} label="iOS" /> : null}
-      {app.androidUrl ? <OriginLink href={app.androidUrl} label="Android" /> : null}
-      {app.fdroidUrl ? <OriginLink href={app.fdroidUrl} label="F-Droid" /> : null}
-      {app.obtainiumUrl ? <OriginLink href={app.obtainiumUrl} label="Obtainium" /> : null}
-    </div>
-  </article>;
-}
-
-function PlatformIconList({ platforms, setPlatform }: { platforms: RecommendedPlatformId[]; setPlatform: (platform: RecommendedPlatformFilterId) => void }): ReactElement | null {
-  if (platforms.length === 0) {
-    return null;
-  }
-
-  return <span className="os-icon-list platform-icon-list">{platforms.map((platform) => <PlatformIcon id={platform} key={platform} setPlatform={setPlatform} />)}</span>;
-}
-
-function PlatformIcon({ id, setPlatform }: { id: RecommendedPlatformId; setPlatform: (platform: RecommendedPlatformFilterId) => void }): ReactElement {
-  const iconMap: Record<RecommendedPlatformId, { Icon: IconType; label: string }> = {
-    webapp: { Icon: FiMonitor, label: "WebApp" },
-    saas: { Icon: FiGlobe, label: "SaaS" },
-    cli: { Icon: FiTerminal, label: "CLI" },
-    api: { Icon: FiCode, label: "API" },
-  };
-  const { Icon, label } = iconMap[id];
-
-  return <button aria-label={`Filter by ${label}`} data-tooltip={`Filter by ${label}`} onClick={() => setPlatform(id)} title={`Filter by ${label}`} type="button"><Icon aria-hidden="true" focusable="false" /></button>;
-}
-
-function OsIconList({ os, setOs }: { os: RecommendedOsId[]; setOs: (os: RecommendedOsId) => void }): ReactElement | null {
-  if (os.length === 0) {
-    return null;
-  }
-
-  return <span className="os-icon-list">{os.map((item) => <OsIcon id={item} key={item} setOs={setOs} />)}</span>;
-}
-
-function OsIcon({ id, setOs }: { id: RecommendedOsId; setOs: (os: RecommendedOsId) => void }): ReactElement {
-  const iconMap: Record<RecommendedOsId, { Icon: IconType; label: string }> = {
-    all: { Icon: FiGlobe, label: "Web app" },
-    macos: { Icon: FaApple, label: "macOS" },
-    linux: { Icon: FaLinux, label: "Linux" },
-    windows: { Icon: FaWindows, label: "Windows" },
-    ios: { Icon: SiIos, label: "iOS" },
-    android: { Icon: IoLogoAndroid, label: "Android" },
-  };
-  const { Icon, label } = iconMap[id];
-
-  return <button aria-label={`Filter by ${label}`} data-tooltip={`Filter by ${label}`} onClick={() => setOs(id)} title={`Filter by ${label}`} type="button"><Icon aria-hidden="true" focusable="false" /></button>;
-}
-
-function recommendedAppMatchesFilters(app: RecommendedApp, platform: RecommendedPlatformFilterId, os: RecommendedOsId): boolean {
-  const platformMatches = platform === "all" || app.platforms.includes(platform);
-  const osMatches = os === "all" || app.os.includes(os);
-
-  return platformMatches && osMatches;
+function categoryIsOneOf(category: string, candidates: string[]): boolean {
+  return candidates.includes(category);
 }
 
 function applyManagedPolicyToggles(app: GuiManagedAppSummary, policy: Partial<Record<ManagedPolicyId, boolean>> | undefined): GuiManagedAppSummary {
@@ -512,19 +342,6 @@ function writeManagedPolicyToggles(policy: ManagedPolicyToggleState): void {
   }
 
   window.localStorage.setItem("aidevops-gui-managed-app-policy", JSON.stringify(policy));
-}
-
-function openExternalLink(event: ReactMouseEvent<HTMLAnchorElement>, href: string): void {
-  event.stopPropagation();
-  if (event.defaultPrevented || event.metaKey || event.ctrlKey || event.shiftKey || event.altKey) {
-    return;
-  }
-
-  const opened = window.open(href, "_blank", "noopener,noreferrer");
-  if (opened !== null) {
-    event.preventDefault();
-    opened.opener = null;
-  }
 }
 
 export function InstallationSurface(): ReactElement {
